+++ src/heat.js
@@ -828,7 +828,7 @@
 
             for ( var type in colorRangeValuesForCurrentYear.types ) {
                 if ( colorRangeValuesForCurrentYear.types.hasOwnProperty( type ) ) {
-                    renderControlStatisticsDay( type, rangeLines, colorRangeValuesForCurrentYear.types[ type ], bindingOptions, colorRanges, pixelsPerNumbers );
+                    renderControlStatisticsRangeLine( type, rangeLines, colorRangeValuesForCurrentYear.types[ type ], bindingOptions, colorRanges, pixelsPerNumbers );
 
                     if ( bindingOptions.views.statistics.showColorRangeLabels ) {
                         createElementWithHTML( statisticsRanges, "div", "range-name", type + "+" );
@@ -842,13 +842,22 @@
         }
     }
 
-    function renderControlStatisticsDay( colorRangeMinimum, dayLines, rangeCount, bindingOptions, colorRanges, pixelsPerNumbers ) {
-        var rangeLine = createElement( dayLines, "div", "range-line no-hover" ),
+    function renderControlStatisticsRangeLine( colorRangeMinimum, dayLines, rangeCount, bindingOptions, colorRanges, pixelsPerNumbers ) {
+        var rangeLine = createElement( dayLines, "div", "range-line" ),
             useColorRange = getColorRangeByMinimum( colorRanges, colorRangeMinimum );
 
         rangeLine.style.height = ( rangeCount * pixelsPerNumbers ) + "px";
 
         addToolTip( rangeLine, bindingOptions, rangeCount.toString() );
+
+        if ( isDefinedFunction( bindingOptions.onStatisticClick ) ) {
+            rangeLine.onclick = function() {
+                fireCustomTrigger( bindingOptions.onStatisticClick, useColorRange );
+            };
+
+        } else {
+            addClass( rangeLine, "no-hover" );
+        }
 
         if ( isDefined( useColorRange ) && isHeatMapColorVisible( bindingOptions, useColorRange.id ) ) {
             if ( isDefinedString( useColorRange.statisticsCssClassName ) ) {
@@ -1593,6 +1602,7 @@
         options.onViewSwitch = getDefaultFunction( options.onViewSwitch, null );
         options.onColorRangeTypeToggle = getDefaultFunction( options.onColorRangeTypeToggle, null );
         options.onImport = getDefaultFunction( options.onImport, null );
+        options.onStatisticClick = getDefaultFunction( options.onStatisticClick, null );
 
         return options;
     }